Fix run options logging when not specified
if run trace was not set or was set to none, run options would still be logged to the graph. This would cause a situation in which tensorboard would attempt to show the run statistics but instead get hung looking for metadata.pbtxt, this will now not be logged unless specified.  The names of loss/acc/auc are now improved such that training and val are displayed side-by-side in tensorboard. However, this should be further improved so that they are included on the same graph