multiclass classification is ~"working"...
mnist multiclass classification is running and it is "working" but it's messy. There are a lot of new TODO:s.. The big issue is handing input data and tf records. I will need to add some sanity checks and logic to ensure that the correct input type is used (one hot). Also, the core logic for this classification really boils down to softmax vs sigmoid not binary vs multi.. I think I should re-implement this logic. I "could" use softmax all the time for classification, but I'm not sure this is the route I want to go yet.