Begin setup for initialization from ckpt for transfer learning
This is the first step for setting up being able to preform transfer learning. The plan is to specify a path to the params to load from in model_config (complete) and use this path as default. Then, in hidden config there will be an option for a path (in case weight's are loaded from different files).