Include basic support for logging (prototype)
Logging is now supported. However, it's safe to say best practices are not being followed. Currently a single file `yf_logging.py` creates and returns the logger, eventually, this will be replaced by importing the logging module and using `logging.getLogger(__name__)`. The messages will also need to be added/improved and the print statements will need to be removed. Also, I think a logger for the `yaml_parse` file (which should likely be called `setup` would be useful