main type config renamed to "loss_fn" + include iris placeholder
The main parameter used to determine "the type" of problem is more accurately described by "loss_fn" as the parameter is controlling the loss function that is being used. Additionally, a directory was created for the iris dataset. The next phase will be implementing regression and the iris dataset was selected for is small size, ease of attaining, and popularity