Update datahandling to manage reshaping
Previously some data would be parsed into [1 0 1 0 ... 1 0 1] rather than the expected [[1] [0] [1] [1] ... [1] [0] [1]]. and in mnist the shape would be 1,28,28,1 where each value was [n] enclosed rather than the expected row/col grouping. The updated solution is not perfect, but it is working at the moment.  Future work will involve cleaning up this functionality and documenting the behavior. This allows us to remove the numpy reshape in both the train and eval