Update regression config - All examples running w/o manual intervention
This is a milestone commit in that all the examples are now running by simply selecting the desired example. The config file is becoming too specific/not general.. the MCd["num_classes] logic should be adjusted to correspond to the output dim to prevent mistakes/inconsistencies between the two. The get_parse_type for tf_records has been adjusted to accept capital/lower case letters and also exits with an error related to unsupported types.