Automatically handle rank error w/input to recurrent
recurrent layers require the input to be rank > 3 and this will automatically apply a expand_dims (and squeeze) if it seems necessary. by "seems" I mean if the rank is < 3 and the last dimension is != 1. this allows for adding and removing a dim easily. This may not be perfect, but it does seem to work in this instance. e.g. (None, 256) will become (None, 256, 1) and then will become (None, 256) again.