Return dictionaries from train and eval
By returning dictionaries this allows the YeahML logic to be treated as a "black box" where a config file is passed in (how and what to train+eval) and a dictionary is returned (results from train+eval). This allows a "meta-program" to sit on top of YeahML as an "AutoML" foundation. This library has been started [CrystalBall](https://github.com/JackBurdick/CrystalBall).