Update initializer logic
I'm not thrilled with this logic overall. Initializers are either classes or functions in tf2 and I opted to implement the classes. I am still not sure how this will work with custom initializers and will likely have to revisit this logic -- especially since I am using the `get_config()` to update parameters.