save model
simplify the save logic. Now, a `save/` directory is made with separate entries for params and the model. Currently, the model is only saved if the model is trained and params need to be saved (the loss is < np.inf). 

eventually it would be nice to have logic to save checkpoints (if desired)