Include dataduit example
This is a POC that shows how to use https://github.com/JackBurdick/dataduit with yeahml.

The example is not perfect, but it is working.

The config/data/parse_data was modified to not be required (by wrapping the whole block in a try/pass. this will need to change, but I'm not sure how to best handle this yet.

Additionally, I'll want to have dataduit and yeahml work together a little more seemlessly (especially around the datatype/parsing information).

Nevertheless, this is a really exciting step and what I have been working towards.