build model
this is a fairly big milestone. the example model is being built using the new configs. I've also included a new `label` flag for the layers/data such that we know which layers to ignore when building the graph

assumptions: that the order the start points and endpoints are defined is the order they are requested in the model. this may need to change -- maybe as a dict?