hacky ("working"!) multi-loss
this is an important milestone, that applies optimizations over multiple losses. There is a ton to clean up, but this is worth celebrating (separately, of course, as we're all in covid19 quarantine).

the next step will be to ensure we track all of the losses+metrics correctly as well as group metrics based on their y/pred specifications (if supervised -- I'm only working with supervised at the moment).

we will also need to implement the logic for multiple optimizers (this has started) and how to handle multiple optimizers (sequentially, alternating, etc.)