selecting optimizers and objectives
This is a start, not the final say -- but the issue is that we were sequentially stepping through the optimizers then objectives, but in reality, we will need to be able to select which optimizers and objectives to use.

also, I'm not convinced the optimizers should be driving the outer loop - but I'm not sure what should yet.