include loss scalar
This does plot the loss scalars as expected, but it does not organize the scalars on tensorboard as I might expect. The issue is that all losses are placed under the first sub directory "/loss", but the others are not further separable. not sure if/how this is possible yet