Include notes on gradient functions
tf.functions expect the same size/shape input every time. The issue is that sometimes batches are different sizes (for example, when working with text data). One option is to use the supported padded_batch() with the max_sequence option, but you may not know that size, and if you pick a "large" size, if a sequence is entered that is longer, an error will be thrown. One could perform a preprocessing step to clip/chop/slice to a set seq as a work around and that is similar to what I've done